requirements.txt
alembic==1.12.0
ansi2html==1.8.0
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.8.1
asttokens==2.3.0
async-lru==2.0.4
#attrs==22.2.0
autopep8==2.3.0
Babel==2.12.1
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==4.0.1
beautifulsoup4==4.11.1
black==24.4.2
bleach==6.0.0
blessed==1.19.1
blinker==1.6.2
blis==0.7.9
branca==0.6.0
cachelib==0.9.0
cairocffi==1.6.1
CairoSVG==2.7.1
camelot-py==0.11.0
catalogue==2.0.8
certifi==2022.9.24
cffi==1.15.1
chardet==5.2.0
charset-normalizer==2.1.1
click==8.1.7
click-log==0.4.0
click-plugins==1.1.1
cligj==0.7.2
colorlover==0.3.0
comm==0.1.4
confection==0.0.3
contourpy==1.1.0
cryptography==41.0.3
cssselect2==0.7.0
cycler==0.11.0
cymem==2.0.7
dash==2.13.0
dash-bootstrap-components==1.4.2
dash-bootstrap-templates==1.0.8
dash-core-components==2.0.0
dash-dangerously-set-inner-html==0.0.2
dash-daq==0.5.0
dash-datetimepicker==0.0.7
dash-defer-js-import==0.0.2
dash-down==0.1.2
dash-extensions==1.0.3
dash-html-components==2.0.0
dash-iconify==0.1.2
dash-leaflet==1.0.8
dash-loading-spinners==1.0.0
dash-mantine-components==0.12.1
dash-renderer==1.9.1
dash-svg==0.0.12
dash-table==5.0.0
dataclass-wizard==0.22.2
datatables==0.4.9
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
diskcache==5.6.3
Django==4.2.13
dnspython==2.4.2
docxcompose==1.4.0
docxtpl==0.16.7
dominate==2.8.0
EditorConfig==0.12.3
elastic-transport==8.4.0
elasticsearch==8.9.0
email-validator==1.1.2
eml-parser==1.17.5
enlighten==1.11.1
entrypoints==0.4
et-xmlfile==1.1.0
evdev==1.7.1
exceptiongroup==1.1.3
executing==1.2.0
fake-useragent==1.3.0
fastjsonschema==2.18.0
#Fiona==1.8.19
Flask==2.2.5
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-Caching==2.0.2
Flask-Cors==4.0.0
Flask-DebugToolbar==0.13.1
Flask-Executor==1.0.0
Flask-HTTPAuth==4.8.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Moment==1.0.5
Flask-SQLAlchemy==3.1.1
Flask-Static-Digest==0.4.0
Flask-WTF==1.1.1
folium==0.14.0
fontawesomefree==6.4.2
fonttools==4.42.1
fqdn==1.5.1
#fr-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_lg-3.4.0/fr_core_news_lg-3.4.0-py3-none-any.whl#sha256=b68c2e0472a66287690293cd08b47a7947a1df4ae4f729de1fd48983dd0beaf9
future==0.18.2
#GDAL==3.0.4
#geobuf==1.1.1
#geographiclib==2.0
#geopandas==0.13.2
#geopy==2.4.0
ghostscript==0.7
grasia-dash-components==0.3.4
greenlet==2.0.1
gunicorn==21.2.0
h11==0.14.0
html2text==2024.2.26
html5lib==1.1
humanfriendly==10.0
idna==3.4
image==1.5.33
importlib-metadata==6.8.0
importlib-resources==6.0.1
iniconfig==2.0.0
ipykernel==6.25.2
ipyleaflet==0.17.4
ipython==8.12.3
ipython-genutils==0.2.0
ipython-sql==0.5.0
ipywidgets==8.1.0
isbnlib==3.10.3
iso3166==1.0.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsbeautifier==1.14.9
json5==0.9.14
jsonpointer==2.4
#jsonschema==4.18.0
#jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-contrib-core==0.4.2
jupyter-contrib-nbextensions==0.7.0
jupyter-events==0.7.0
jupyter-highlight-selected-word==0.2.0
jupyter-lsp==2.2.0
jupyter-nbextensions-configurator==0.6.3
jupyter_client==7.4.9
jupyter_core==5.3.2
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
jwt==1.3.1
kaleido==0.2.1
keyboard==0.13.5
kiwisolver==1.4.5
langcodes==3.3.0
libpff-python-ratom==20220304
#libratom==0.7.1
livereload @ https://github.com/nacho00112/python-livereload/archive/refs/heads/master.zip#sha256=08bab1332a2a44e2ab12359eecb20e6f2bec8d50a7230778d43ffd32272a09fa
lxml==4.9.3
mail-parser==3.15.0
Mako==1.2.4
Markdown==3.4.4
MarkupSafe==2.1.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==2.0.4
more-itertools==9.1.0
msal==1.28.0
multiprocess==0.70.15
munch==4.0.0
murmurhash==1.0.9
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.1
notebook==6.5.6
notebook_shim==0.2.3
numpy==1.23.5
olefile==0.46
opencv-python==4.9.0.80
openpyxl==3.1.2
#openrouteservice==2.3.3
#osmnx==1.9.3
outcome==1.2.0
overrides==7.4.0
packaging==24.1
pandas==2.0.3
pandas-datareader==0.10.0
pandas-geojson==1.2.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.12.1
pathy==0.10.0
pbr==5.11.0
pdfminer==20191125
pdfminer.six==20221105
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
plotly==5.16.1
pluggy==1.3.0
ppt2txt==0.1.0
prefixed==0.5.0
preshed==3.0.8
prettytable==3.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==3.20.0
psutil==5.9.4
psycopg2==2.9.7
psycopg2-binary==2.9.7
ptyprocess==0.7.0
publicsuffixlist==0.10.0.20240515
pure-eval==0.2.2
py-moneyed==0.8.0
pyasn1==0.6.0
pycodestyle==2.12.0
pycountry==22.3.5
pycparser==2.21
pycryptodome==3.20.0
pydantic==1.10.2
pygal==3.0.4
Pygments==2.16.1
PyJWT==2.8.0
pynput==1.7.7
pyodbc==4.0.39
pyparsing==3.0.9
pypdf==3.17.0
PyPDF2==3.0.1
pyproj==3.5.0
pyrsistent==0.19.2
PySimpleGUI==5.0.5
PySocks==1.7.1
pytesseract==0.3.10
pytest==7.4.1
python-box==6.1.0
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-json-logger==2.0.7
python-xlib==0.33
pytz==2020.1
PyVirtualDisplay==3.0
PyYAML==6.0.1
pyzmq==24.0.0
qgrid==1.3.1
qtconsole==5.4.4
QtPy==2.4.0
randomtimestamp==2.2
referencing==0.30.2
requests==2.28.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.0
rsa==4.9
scapy==2.5.0
schwifty==2020.9.0
scikit-learn==1.3.2
scipy==1.10.1
Send2Trash==1.8.2
shapely==2.0.4
simple-date==0.5.0
simplejson==3.19.2
six==1.16.0
smart-open==5.2.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.4.3
spacy-legacy==3.0.10
spacy-loggers==1.0.3
SQLAlchemy==2.0.16
SQLAlchemy-serializer==1.4.12
sqlparse==0.5.0
srsly==2.4.5
stack-data==0.6.2
stem==1.8.2
striprtf==0.0.22
tabulate==0.9.0
tenacity==8.2.3
terminado==0.17.1
thinc==8.1.5
threadpoolctl==3.2.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.3
torrequest==0.1.0
tqdm==4.64.1
traitlets==5.11.2
traittypes==0.2.1
treelib==1.6.1
trio==0.22.2
trio-websocket==0.11.1
typer==0.7.0
typing_extensions==4.4.0
tzdata==2023.3
tzlocal==5.2
unicode==2.9
Unidecode==1.3.7
uri-template==1.3.0
urllib3==1.26.13
uWSGI==2.0.22
visdcc==0.0.50
visitor==0.1.3
wasabi==0.10.1
wcwidth==0.2.5
webcolors==1.13
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.6.2
Werkzeug==2.2.3
wget==3.2
widgetsnbextension==4.0.8
wordcloud==1.9.2
wsproto==1.2.0
WTForms==3.0.1
xyzservices==2023.7.0
zipp==3.16.2
